+++ docs/changelog.md
@@ -25,6 +25,7 @@ This changelog follows the rules of [Keep a Changelog](http://keepachangelog.com
 
 - A subtle pulse animation to the drop indicator in the menu editor's preview.
 - The possibility to lock item positions in the menu preview of the editor. Locked items cannot be reordered but moved to a fixed angle instead.
+- The possibility to close the menu with the right mouse button.
 
 #### Changed